+++ libs/evm-protocols/src/common-protocol/utils.ts
@@ -1,4 +1,6 @@
-import Web3, { AbiInput, TransactionReceipt } from 'web3';
+import Web3, { AbiInput, TransactionReceipt, Web3 as Web3Type } from 'web3';
+
+export type EvmClientType = Web3Type;
 
 export const calculateVoteWeight = (
   balance: string, // should be in wei
@@ -27,24 +29,64 @@ export const getAddressFromSignedMessage = (
   return web3.eth.accounts.recover(message, signature);
 };
 
+export const getBlock = async ({
+  evmClient,
+  rpc,
+  blockHash,
+}: {
+  evmClient?: EvmClientType;
+  rpc: string;
+  blockHash: string;
+}) => {
+  const web3 = evmClient || new Web3(rpc);
+  return {
+    block: await web3.eth.getBlock(blockHash),
+    evmClient: web3,
+  };
+};
+
 export const getTransactionReceipt = async ({
+  evmClient,
   rpc,
   txHash,
 }: {
+  evmClient?: EvmClientType;
   rpc: string;
   txHash: string;
-}): Promise<TransactionReceipt> => {
-  const web3 = new Web3(rpc);
-  return await web3.eth.getTransactionReceipt(txHash);
+}): Promise<{
+  txReceipt: TransactionReceipt;
+  evmClient: EvmClientType;
+}> => {
+  const web3 = evmClient || new Web3(rpc);
+  return {
+    txReceipt: await web3.eth.getTransactionReceipt(txHash),
+    evmClient: web3,
+  };
 };
 
 export const decodeParameters = ({
+  evmClient,
   abiInput,
   data,
 }: {
+  evmClient?: EvmClientType;
   abiInput: AbiInput[];
   data: string;
 }) => {
-  const web3 = new Web3();
+  const web3 = evmClient || new Web3();
   return web3.eth.abi.decodeParameters(abiInput, data);
+};
+
+export const createPrivateEvmClient = ({
+  rpc,
+  privateKey,
+}: {
+  rpc: string;
+  privateKey: string;
+}): Web3 => {
+  const web3 = new Web3(rpc);
+  const account = web3.eth.accounts.privateKeyToAccount(privateKey);
+  web3.eth.accounts.wallet.add(account);
+  web3.eth.defaultAccount = account.address;
+  return web3;
 };
