+++ libs/model/src/policies/handleLaunchpadTrade.ts
@@ -1,11 +1,15 @@
 import { logger } from '@hicommonwealth/core';
-import { commonProtocol as cp } from '@hicommonwealth/evm-protocols';
+import {
+  commonProtocol as cp,
+  getLaunchpadToken,
+  transferLaunchpadLiquidityToUniswap,
+} from '@hicommonwealth/evm-protocols';
+import { config } from '@hicommonwealth/model';
 import { chainEvents, events } from '@hicommonwealth/schemas';
 import { BigNumber } from 'ethers';
 import Web3 from 'web3';
 import { z } from 'zod';
 import { models } from '../database';
-import { commonProtocol } from '../services';
 import { chainNodeMustExist } from './utils';
 
 const log = logger(import.meta);
@@ -77,17 +81,18 @@ export async function handleLaunchpadTrade(
     BigNumber.from(floatingSupply).toBigInt() ===
       BigInt(token.launchpad_liquidity)
   ) {
-    const onChainTokenData = await commonProtocol.launchpadHelpers.getToken({
+    const onChainTokenData = await getLaunchpadToken({
       rpc: chainNode.private_url!,
       tokenAddress,
       lpBondingCurveAddress,
     });
 
     if (!onChainTokenData.funded) {
-      await commonProtocol.launchpadHelpers.transferLiquidityToUniswap({
+      await transferLaunchpadLiquidityToUniswap({
         rpc: chainNode.private_url!,
         tokenAddress,
         lpBondingCurveAddress,
+        privateKey: config.WEB3.PRIVATE_KEY,
       });
     }
