+++ libs/model/src/token/CreateLaunchpadTrade.command.ts
@@ -1,10 +1,12 @@
 import { Command, InvalidState } from '@hicommonwealth/core';
-import { commonProtocol } from '@hicommonwealth/evm-protocols';
+import {
+  commonProtocol,
+  getLaunchpadTradeTransaction,
+} from '@hicommonwealth/evm-protocols';
 import * as schemas from '@hicommonwealth/schemas';
 import z from 'zod';
 import { models } from '../database';
 import { mustExist } from '../middleware/guards';
-import { getLaunchpadTradeTransaction } from '../services/commonProtocol/launchpadHelpers';
 
 const launchpadEthChainIds = Object.values(
   commonProtocol.factoryContracts,
